* fix/misc:
  Fixed issue with migration not working correctly for legacy feeding objects
  Bumped version code
  Added call for migration
  Added back deprecated models for migration
  Updated readmes
  Changed type of amount for additive
  Added instance state saving for main activity

+++ app/src/main/java/me/anon/grow/fragment/AddAdditiveDialogFragment.java
@@ -68,7 +68,7 @@ public class AddAdditiveDialogFragment extends DialogFragment
 					Additive additive = new Additive();
 
 					String desc = TextUtils.isEmpty(description.getText()) ? null : description.getText().toString();
-					Integer amt = TextUtils.isEmpty(amount.getText()) ? null : Integer.valueOf(amount.getText().toString());
+					Double amt = TextUtils.isEmpty(amount.getText()) ? null : Double.valueOf(amount.getText().toString());
 
 					additive.setDescription(desc);
 					additive.setAmount(amt);
